+++ lib/pay/braintree/webhooks/subscription_trial_ended.rb
@@ -3,13 +3,7 @@
 module Pay
   module Braintree
     module Webhooks
-      class SubscriptionTrialEnded
-        def call(event)
-          subscription = event.subscription
-          return if subscription.nil?
-
-          Pay::Braintree::Subscription.sync(subscription.id)
-        end
+      class SubscriptionTrialEnded < Subscription
       end
     end
   end